+++ src/clauses/Match.ts
@@ -23,6 +23,7 @@ import type { QuantifiedPath } from "../pattern/quantified-patterns/QuantifiedPa
 import { compileCypherIfExists } from "../utils/compile-cypher-if-exists";
 import { Clause } from "./Clause";
 import { WithPathAssign } from "./mixins/WithPathAssign";
+import { WithCall } from "./mixins/clauses/WithCall";
 import { WithCallProcedure } from "./mixins/clauses/WithCallProcedure";
 import { WithCreate } from "./mixins/clauses/WithCreate";
 import { WithFinish } from "./mixins/clauses/WithFinish";
@@ -48,7 +49,8 @@ export interface Match
         WithCreate,
         WithMerge,
         WithFinish,
-        WithCallProcedure {}
+        WithCallProcedure,
+        WithCall {}
 
 /**
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/clauses/match/)
@@ -66,7 +68,8 @@ export interface Match
     WithCreate,
     WithMerge,
     WithFinish,
-    WithCallProcedure
+    WithCallProcedure,
+    WithCall
 )
 export class Match extends Clause {
     private pattern: Pattern | QuantifiedPath;