+++ src/clauses/With.ts
@@ -23,6 +23,7 @@ import type { Variable } from "../references/Variable";
 import type { Expr } from "../types";
 import { compileCypherIfExists } from "../utils/compile-cypher-if-exists";
 import { Clause } from "./Clause";
+import { WithCall } from "./mixins/clauses/WithCall";
 import { WithCallProcedure } from "./mixins/clauses/WithCallProcedure";
 import { WithCreate } from "./mixins/clauses/WithCreate";
 import { WithMatch } from "./mixins/clauses/WithMatch";
@@ -47,13 +48,25 @@ export interface With
         WithUnwind,
         WithCreate,
         WithMerge,
-        WithCallProcedure {}
+        WithCallProcedure,
+        WithCall {}
 
 /**
  * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/clauses/with/)
  * @category Clauses
  */
-@mixin(WithOrder, WithReturn, WithWhere, WithDelete, WithMatch, WithUnwind, WithCreate, WithMerge, WithCallProcedure)
+@mixin(
+    WithOrder,
+    WithReturn,
+    WithWhere,
+    WithDelete,
+    WithMatch,
+    WithUnwind,
+    WithCreate,
+    WithMerge,
+    WithCallProcedure,
+    WithCall
+)
 export class With extends Clause {
     private projection: Projection;
     private isDistinct = false;